test for LinearRegression
Actually doing an lr_model.fit won't work because the train and test index will need to have the same number of indexes but in our code its a 1:4 ration of indexes. Reshaping may need to be considered or if another approach is used it may mean we will have missing data. Food for thought.